+++ setup/docker/entrypoint.bash
@@ -1,5 +1,0 @@
-#!/usr/bin/env bash
-cd /srv/enketo-express
-python setup/docker/create_config.py
-grunt
-npm start